Move some functions from packager to rewriter
